requirements.txt
asgiref==3.2.10
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1
django-cors-headers==3.4.0
django-import-export==2.3.0
django-rest-framework==0.1.0
djangorestframework==3.11.1
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
psycopg2==2.8.5
psycopg2-binary==2.8.5
pytz==2020.1
PyYAML==5.3.1
sqlparse==0.3.1
tablib==2.0.0
xlrd==1.2.0
xlwt==1.3.0
